requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
email-validator==2.1.0
imapclient==3.0.1
email-reply-parser==0.5.11
plotly==5.17.0
wordcloud==1.9.2
textblob==0.17.1
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
streamlit-extras==0.3.0
